+++ test/OpenTelemetry.Instrumentation.AspNetCore.Tests/RouteTests/RoutingTestCases.cs
@@ -6,7 +6,6 @@
 using System.Reflection;
 using System.Text.Json;
 using System.Text.Json.Serialization;
-using RouteTests.TestApplication;
 
 namespace RouteTests;
 
@@ -42,30 +41,5 @@ public static class RoutingTestCases
         }
 
         return result;
-    }
-
-    public class TestCase
-    {
-        public string Name { get; set; } = string.Empty;
-
-        public int? MinimumDotnetVersion { get; set; }
-
-        public TestApplicationScenario TestApplicationScenario { get; set; }
-
-        public string? HttpMethod { get; set; }
-
-        public string Path { get; set; } = string.Empty;
-
-        public int ExpectedStatusCode { get; set; }
-
-        public string? ExpectedHttpRoute { get; set; }
-
-        public string? CurrentHttpRoute { get; set; }
-
-        public override string ToString()
-        {
-            // This is used by Visual Studio's test runner to identify the test case.
-            return $"{this.TestApplicationScenario}: {this.Name}";
-        }
     }
 }
