+++ test/OpenTelemetry.Instrumentation.AspNetCore.Tests/RouteTests/TestApplication/PageActionDescriptorInfo.cs
@@ -1,0 +1,26 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+#nullable enable
+using System.Text.Json.Serialization;
+
+namespace RouteTests.TestApplication;
+
+public class PageActionDescriptorInfo
+{
+    public PageActionDescriptorInfo()
+    {
+    }
+
+    public PageActionDescriptorInfo(string relativePath, string viewEnginePath)
+    {
+        this.PageActionDescriptorRelativePath = relativePath;
+        this.PageActionDescriptorViewEnginePath = viewEnginePath;
+    }
+
+    [JsonPropertyName("RelativePath")]
+    public string PageActionDescriptorRelativePath { get; set; } = string.Empty;
+
+    [JsonPropertyName("ViewEnginePath")]
+    public string PageActionDescriptorViewEnginePath { get; set; } = string.Empty;
+}
