+++ test/OpenTelemetry.Instrumentation.AspNetCore.Tests/RouteTests/TestApplication/RouteInfo.cs
@@ -9,8 +9,6 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.Metadata;
 #endif
 using Microsoft.AspNetCore.Mvc.Abstractions;
-using Microsoft.AspNetCore.Mvc.Controllers;
-using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.AspNetCore.Routing;
 
 namespace RouteTests.TestApplication;
@@ -56,83 +54,5 @@ public class RouteInfo
         {
             this.ActionDescriptor = new ActionDescriptorInfo(actionDescriptor);
         }
-    }
-
-    public class ActionDescriptorInfo
-    {
-        public ActionDescriptorInfo()
-        {
-        }
-
-        public ActionDescriptorInfo(ActionDescriptor actionDescriptor)
-        {
-            this.AttributeRouteInfo = actionDescriptor.AttributeRouteInfo?.Template;
-
-            this.ActionParameters = new List<string>();
-            foreach (var item in actionDescriptor.Parameters)
-            {
-                this.ActionParameters.Add(item.Name);
-            }
-
-            if (actionDescriptor is PageActionDescriptor pad)
-            {
-                this.PageActionDescriptorSummary = new PageActionDescriptorInfo(pad.RelativePath, pad.ViewEnginePath);
-            }
-
-            if (actionDescriptor is ControllerActionDescriptor cad)
-            {
-                this.ControllerActionDescriptorSummary = new ControllerActionDescriptorInfo(cad.ControllerName, cad.ActionName);
-            }
-        }
-
-        [JsonPropertyName("AttributeRouteInfo.Template")]
-        public string? AttributeRouteInfo { get; set; }
-
-        [JsonPropertyName("Parameters")]
-        public IList<string>? ActionParameters { get; set; }
-
-        [JsonPropertyName("ControllerActionDescriptor")]
-        public ControllerActionDescriptorInfo? ControllerActionDescriptorSummary { get; set; }
-
-        [JsonPropertyName("PageActionDescriptor")]
-        public PageActionDescriptorInfo? PageActionDescriptorSummary { get; set; }
-    }
-
-    public class ControllerActionDescriptorInfo
-    {
-        public ControllerActionDescriptorInfo()
-        {
-        }
-
-        public ControllerActionDescriptorInfo(string controllerName, string actionName)
-        {
-            this.ControllerActionDescriptorControllerName = controllerName;
-            this.ControllerActionDescriptorActionName = actionName;
-        }
-
-        [JsonPropertyName("ControllerName")]
-        public string ControllerActionDescriptorControllerName { get; set; } = string.Empty;
-
-        [JsonPropertyName("ActionName")]
-        public string ControllerActionDescriptorActionName { get; set; } = string.Empty;
-    }
-
-    public class PageActionDescriptorInfo
-    {
-        public PageActionDescriptorInfo()
-        {
-        }
-
-        public PageActionDescriptorInfo(string relativePath, string viewEnginePath)
-        {
-            this.PageActionDescriptorRelativePath = relativePath;
-            this.PageActionDescriptorViewEnginePath = viewEnginePath;
-        }
-
-        [JsonPropertyName("RelativePath")]
-        public string PageActionDescriptorRelativePath { get; set; } = string.Empty;
-
-        [JsonPropertyName("ViewEnginePath")]
-        public string PageActionDescriptorViewEnginePath { get; set; } = string.Empty;
     }
 }
